home-again-api/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "MIDS Capstone - Home Again API"
authors = ["Jonathan Mulieri <jmulieri@ischool.berkeley.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.10"
fastapi = "^0.103.1"
joblib = "^1.3.2"
scikit-learn = "^1.3.1"
scipy = "^1.11.3"
fastapi-cache2 = {version = "0.1.9", extras = ["redis"]}
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.36"
python-dotenv = "^1.0.1"
#psycopg2 = "^2.9.10"
psycopg2-binary = "^2.9.10"
pendulum = "2.0.3"
pandas = "^2.2.3"
pyjwt = "^2.9.0"
cryptography = "^43.0.3"
boto3 = "^1.35.68"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.1"
httpx = "^0.24.1"
ruff = "^0.0.287"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
